requirements.txt
django==3.0.6
django-model-utils==4.0.0
django-cors-headers==3.2.1
djangorestframework==3.11.0
pytest==5.4.1
pytest-django==3.9.0
factory_boy==2.12.0
drf-yasg==1.17.1
django-environ==0.4.5
